artifacts/ProjectUltronv1/project_ultron/requirements.txt
ollama
openai
pygame
piper-tts